AM-1273: Fixing the service bus connection in preview.
