+++ build.gradle
@@ -311,7 +311,7 @@ dependencies {
     implementation group: 'com.azure', name: 'azure-core', version: '1.16.0'
     compile group: 'com.microsoft.azure', name: 'azure-servicebus', version: '3.6.3'
     implementation group: 'com.azure', name: 'azure-messaging-servicebus', version: '7.2.2'
-    implementation group: 'com.google.code.gson', name: 'gson', version: '2.8.6'
+    //implementation group: 'com.google.code.gson', name: 'gson', version: '2.8.6'
 
     compile group: 'io.projectreactor', name: 'reactor-core', version: '3.4.0'
     compile group: 'commons-cli', name: 'commons-cli', version: '1.4'
@@ -321,7 +321,6 @@ dependencies {
     compile group: 'org.springframework.hateoas', name: 'spring-hateoas', version: '1.2.2'
     compile group: 'com.google.guava', name: 'guava', version: '30.1-jre'
     compile group: 'org.springframework.retry', name: 'spring-retry', version: '1.3.0'
-    compile group: 'org.amqphub.spring', name: 'amqp-10-jms-spring-boot-starter', version: '2.3.9'
     compile group: 'com.github.ben-manes.caffeine', name: 'caffeine', version: '2.8.5'
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-cache', version: '2.4.1'
