+++ src/functionalTest/java/uk/gov/hmcts/reform/orgrolemapping/befta/OrgRoleMappingAmTestAutomationAdapter.java
@@ -4,7 +4,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.hmcts.befta.DefaultTestAutomationAdapter;
 import uk.gov.hmcts.befta.player.BackEndFunctionalTestScenarioContext;
-import uk.gov.hmcts.befta.util.EnvironmentVariableUtils;
 
 import java.util.UUID;
 
@@ -33,7 +32,7 @@ public class OrgRoleMappingAmTestAutomationAdapter extends DefaultTestAutomation
                 }
                 logger.info("The nap is complete.");
                 return null;
-            case ("generateServiceBusToken"):
+            /*case ("generateServiceBusToken"):
                 try {
                     return FunctionalTestUtils.getSaSToken("sb://"
                                     + EnvironmentVariableUtils.getRequiredVariable("AMQP_HOST"),
@@ -42,7 +41,7 @@ public class OrgRoleMappingAmTestAutomationAdapter extends DefaultTestAutomation
                 } catch (Exception e) {
                     logger.warn(e.getMessage());
                 }
-                return null;
+                return null;*/
             /*case("getUserIdFromFile"):
                 return FunctionalTestUtils.getUserIdFromFile("SingleUserId.json");*/
             default:
